+++ example/android/app/src/main/kotlin/com/example/heic_to_png_jpg_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.heic_to_png_jpg_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()