+++ lib/src/platform_interface.dart
@@ -1,10 +1,7 @@
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'dart:typed_data';
 
-enum ImageFormat {
-  jpg,
-  png,
-}
+import 'package:heic_to_png_jpg/src/image_format.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 abstract class HeicToImagePlatform extends PlatformInterface {
   HeicToImagePlatform() : super(token: _token);
@@ -13,8 +10,7 @@ abstract class HeicToImagePlatform extends PlatformInterface {
   static HeicToImagePlatform? _instance;
 
   static HeicToImagePlatform get instance {
-    return _instance ??=
-        throw Exception('HeicToImagePlatform instance not set');
+    return _instance ??= throw Exception('HeicToImagePlatform instance not set');
   }
 
   static set instance(HeicToImagePlatform? instance) {
@@ -29,5 +25,8 @@ abstract class HeicToImagePlatform extends PlatformInterface {
     ImageFormat format = ImageFormat.jpg,
     int quality = 80,
     int? maxWidth,
+
+    /// To override the default libheif js cdn url.
+    String? libheifJsUrl,
   });
 }